sparse: Use int64_t for file length
Because the image to be flashed might exceed 2G, even 4G, which will
be out of the range of int or uint type. So we need to use int64_t to
indicate the file length.

Reviewed-on: http://gerrit.mot.com/687094
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Connie Zhao <czhao1@motorola.com>

+++ libsparse/output_file.h
@@ -27,15 +27,15 @@ struct output_file;
 
 struct output_file *output_file_open_fd(int fd, unsigned int block_size, int64_t len,
 		int gz, int sparse, int chunks, int crc);
-struct output_file *output_file_open_callback(int (*write)(void *, const void *, int),
+struct output_file *output_file_open_callback(int (*write)(void *, const void *, int64_t),
 		void *priv, unsigned int block_size, int64_t len, int gz, int sparse,
 		int chunks, int crc);
-int write_data_chunk(struct output_file *out, unsigned int len, void *data);
-int write_fill_chunk(struct output_file *out, unsigned int len,
+int write_data_chunk(struct output_file *out, int64_t len, void *data);
+int write_fill_chunk(struct output_file *out, int64_t len,
 		uint32_t fill_val);
-int write_file_chunk(struct output_file *out, unsigned int len,
+int write_file_chunk(struct output_file *out, int64_t len,
 		const char *file, int64_t offset);
-int write_fd_chunk(struct output_file *out, unsigned int len,
+int write_fd_chunk(struct output_file *out, int64_t len,
 		int fd, int64_t offset);
 int write_skip_chunk(struct output_file *out, int64_t len);
 void output_file_close(struct output_file *out);
